+++ pdf/lib/widgets.dart
@@ -33,6 +33,7 @@ part 'widgets/font.dart';
 part 'widgets/geometry.dart';
 part 'widgets/grid_view.dart';
 part 'widgets/image.dart';
+part 'widgets/multi_page.dart';
 part 'widgets/placeholders.dart';
 part 'widgets/stack.dart';
 part 'widgets/table.dart';