+++ pdf/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add available dimensions for PdfPageFormat
 * Add Document properties
 * Add Page.orientation to force landscape or portrait
+* Improve MultiPage Widget
 
 # 1.3.3
 * Fix a bug with the RichText Widget